+++ instances/stop_start_scheduler/CHANGELOG.md
@@ -1,5 +1,10 @@
 Stop/Start Scheduler Policy Changelog
 
+v1.1
+-----
+- Do not try to stop locked instances
+- Ignore errors on servers with ephemeral stores
+
 v1.0
 -----
 - initial release